backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic[email]>=2.10.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
alembic>=1.13.1
openai==0.27.8
PyPDF2>=3.0.1
pypdfium2>=4.0.0
Pillow>=10.0.0
numpy>=1.24.3
selenium>=4.15.0
webdriver-manager>=4.0.1
